backend/requirements.txt
fastapi
uvicorn
pillow
torch
ultralytics
python-multipart
